zmiana typów w init GameObject

+++ src/shared.py
@@ -99,7 +99,7 @@ Metody/gettery:
 - center - zwraca współżędne środka w postaci wektora (x, y)
 """
 class GameObject:
-    def __init__(self, x: int, y: int, width: int, height: int, object_type: ObjectType, velocity: Vector2 | None = None, angular_velocity: float = 0, mass: float = 0, gravity: float = 0, inertia: float = 0, img: pygame.Surface | None = None):
+    def __init__(self, x: float, y: float, width: float, height: float, object_type: ObjectType, velocity: Vector2 | None = None, angular_velocity: float = 0, mass: float = 0, gravity: float = 0, inertia: float = 0, img: pygame.Surface | None = None):
         self.x: float= x
         self.y: float= y
         self.width: float = width
